Rename Diego's class to DPSGP_gpytorch and eliminate temp files
